[eclipse/xtext#1679] converted more Xtend code to Java
