+++ org.eclipse.xtext.ide/src/org/eclipse/xtext/ide/editor/bracketmatching/IBracePairProvider.java
@@ -1,26 +1,29 @@
 /*******************************************************************************
- * Copyright (c) 2015 itemis AG (http://www.itemis.eu) and others.
+ * Copyright (c) 2015, 2020 itemis AG (http://www.itemis.eu) and others.
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License 2.0 which is available at
  * http://www.eclipse.org/legal/epl-2.0.
  *
  * SPDX-License-Identifier: EPL-2.0
  *******************************************************************************/
-package org.eclipse.xtext.ide.editor.bracketmatching
+package org.eclipse.xtext.ide.editor.bracketmatching;
 
-import com.google.inject.Singleton
+import java.util.Set;
 
 /**
+ * <p>
+ * Provides a set of definitions for brace pairs that need to be matched when
+ * editing code.
+ * </p> 
+ * 
  * @author kosyakov - Initial contribution and API
  */
-@Singleton
-class XtextBracePairProvider extends DefaultBracePairProvider {
-	new() {
-		super(#{
-			new BracePair(":", ";", true),
-			new BracePair("(", ")", false),
-			new BracePair("{", "}", true),
-			new BracePair("[", "]", false)
-		})
-	}
+public interface IBracePairProvider {
+
+	/**
+	 * Returns a set of definitions for brace pairs that need to be matched when
+	 * editing code.
+	 */
+	Set<BracePair> getPairs();
+
 }
