+++ org.eclipse.xtext.ide/src/org/eclipse/xtext/ide/editor/bracketmatching/XtextBracePairProvider.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2015 itemis AG (http://www.itemis.eu) and others.
+ * Copyright (c) 2015, 2020 itemis AG (http://www.itemis.eu) and others.
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License 2.0 which is available at
  * http://www.eclipse.org/legal/epl-2.0.
@@ -18,10 +18,9 @@ import org.eclipse.xtext.xbase.lib.CollectionLiterals;
  * @author kosyakov - Initial contribution and API
  */
 @Singleton
-@SuppressWarnings("all")
 public class XtextBracePairProvider extends DefaultBracePairProvider {
-  public XtextBracePairProvider() {
-    super(
-      Collections.<BracePair>unmodifiableSet(CollectionLiterals.<BracePair>newHashSet(new BracePair(":", ";", true), new BracePair("(", ")", false), new BracePair("{", "}", true), new BracePair("[", "]", false))));
-  }
+	public XtextBracePairProvider() {
+		super(Collections.unmodifiableSet(CollectionLiterals.newHashSet(new BracePair(":", ";", true),
+				new BracePair("(", ")", false), new BracePair("{", "}", true), new BracePair("[", "]", false))));
+	}
 }
