+++ org.eclipse.xtext.ide/src/org/eclipse/xtext/ide/server/codelens/ICodeLensService.java
@@ -1,0 +1,28 @@
+/*******************************************************************************
+ * Copyright (c) 2017, 2020 TypeFox GmbH (http://www.typefox.io) and others.
+ * This program and the accompanying materials are made available under the
+ * terms of the Eclipse Public License 2.0 which is available at
+ * http://www.eclipse.org/legal/epl-2.0.
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *******************************************************************************/
+package org.eclipse.xtext.ide.server.codelens;
+
+import java.util.List;
+import org.eclipse.lsp4j.CodeLens;
+import org.eclipse.lsp4j.CodeLensParams;
+import org.eclipse.xtext.ide.server.Document;
+import org.eclipse.xtext.resource.XtextResource;
+import org.eclipse.xtext.util.CancelIndicator;
+
+/**
+ * @author Sven Efftinge - Initial contribution and API
+ */
+public interface ICodeLensService {
+	
+	/**
+	 * Compute code lenses for the given context.
+	 */
+	List<? extends CodeLens> computeCodeLenses(Document document, XtextResource resource, CodeLensParams params, CancelIndicator indicator);
+	
+}
