--- org.eclipse.xtext.ide/xtend-gen/org/eclipse/xtext/ide/server/ICapabilitiesContributor.java
@@ -1,23 +1,0 @@
-/**
- * Copyright (c) 2017 TypeFox GmbH (http://www.typefox.io) and others.
- * This program and the accompanying materials are made available under the
- * terms of the Eclipse Public License 2.0 which is available at
- * http://www.eclipse.org/legal/epl-2.0.
- * 
- * SPDX-License-Identifier: EPL-2.0
- */
-package org.eclipse.xtext.ide.server;
-
-import org.eclipse.lsp4j.InitializeParams;
-import org.eclipse.lsp4j.ServerCapabilities;
-
-/**
- * @author Sven Efftinge - Initial contribution and API
- */
-@SuppressWarnings("all")
-public interface ICapabilitiesContributor {
-  /**
-   * Allows an individual language to contribute to and overwrite properties in the server's capabilities.
-   */
-  ServerCapabilities contribute(final ServerCapabilities capabilities, final InitializeParams params);
-}
